- map拆分地图操作至map_operations
