+++ utils/config/map_move.py
@@ -1,0 +1,10 @@
+MAP_MOVE_NAV_DATA = {
+    "16:9": {
+        "directions": {
+            "down": (250, 900, 250, 300),
+            "left": (250, 900, 850, 900),
+            "up": (1330, 200, 1330, 800),
+            "right": (1330, 200, 730, 200),
+        }
+    }
+}
