+++ utils/map_utils/map_selector.py
@@ -1,8 +1,8 @@
 import questionary
-from utils.config import ConfigurationManager
+from utils.config.config import ConfigurationManager
 from utils.log import log
 from utils.time_utils import TimeUtils
-from utils.map_info import MapInfo
+from utils.map_utils.map_info import MapInfo
 from utils.setting import Setting
 
 cfg = ConfigurationManager()
